navbar updates
* fixed to top of page
* shrunk width to account for grid size
* to align on design, added an empty menu for extended black bar
